Mining using player_id
Whenever an attempt at mining and ore node is performed, player_id of the miner is saved in a dictionary together with total number of attempts by this player.
If the number of hits by a player reaches the ore nodes' HITS_HP property, it is mined.
The dictionary for player storage is reset after 10 seconds of non-interaction or if the node is successfully mined.